refactor: alpine_test (#4)
* Only match draft-*.txt and rfc*.txt

* Lint first

* Don't look at HEAD for the default branch except as a last resort

* Use a bearer token perhaps

* Add a token input

* Use the token parameter

* Damned quoting

* No need for a header on the first set

* Back out whitespace lint

* Remove errant 'let'

* Someone insists that .py is helpful

* Remove raw, which no longer works

* Page numbers are the past

This causes drafts to be rendered in the same way as RFCs, by default.
It opens up the possibility of using tools *other than* rfcdiff for
examining changes between versions.  As textual output is often easier
to handle than HTML when it comes to diff tools, this is a strict
improvement.

This will mean that submissions look different, if the text versions are
submitted.  That is unlikely.  What is more likely is that the submitted
version will be paginated, as that is the xml2rfc default.  This means
that diffs for editors copy (and branches) with the submitted versions
will look worse.  rfcdiff can handle that well enough; other tools
won't be any more usable for that case than they were before.

Overall, this seems to be a strict win.  There might be complaints (I
remain impressed and horrified, but unsurprised, at the reaction to the
lack of page numbers in published RFCs).  That's fine.  I believe that
most people using this toolchain use the HTML rendering primarily, as
that remains more accessible than text.

I am NOT adding a configuration option to disable this, but might be
convinced to.

Closes #251.

* Refinements

* Check for bad tags

This prints a warning when tags aren't good.  I've also modified the
pre-push hook to match.  Of course, hooks are nasty so I don't install
them automatically `make update` does that.

Closes #253.

* Make it black

* The whole thing with .targets.mk is whacky

When you try to upload, make is unable to resolve dependencies that
depend on .targets.mk.  This causes `make upload` to hit the fallback
too easily.  This works around that manually in the most hacky way
imaginable.

Includes a fix to CI for black.

* Downgrade default-branch failure to a warning

* Switch the locale in images

Closes #254.

* Add --name-only

* Add a lint for the default branch

Unfortunately, when you newly create a repository, it doesn't get a
reference for refs/remotes/origin/HEAD.  We should lint for that and
then provide a tool to fix it.

* Third-level headings were too small

* Increase line height; add explicit padding for figures

* easy to miss, hard to unsee

* Build twice a month instead

* Fix a11y for pilcrows

This makes them visible always, but not emphasized so that they don't
distract sighted readers.  This means that screen readers will have this
bump always, but the feedback I received indicated that this was
something that people would prefer to have access to, even if it adds a
little extra noise.

* Remove side-scroll hack for wide tables

The use of inline-block breaks some a11y tools.  The right answer for
overflowing tables is to not overflow.  That is, include less
unwrappable text.

* Dark mode

* Darker

* Fix scrollbars

* Update what setup does

* Test the alpine image

* Cache .targets.mk and save some time

Co-authored-by: Martin Thomson <mt@lowentropy.net>