refactor: alpine (#5)
* First cut at an action

* Updates to build scripts

* More actions

* Fixup make action

* Update comment

* Fix quoting

* Remove cache from docker image

* Add XML2RFC_REFCACHEDIR

* Suppress that error

* Make sure to use XML2RFC_REFCACHEDIR

* Refine shared cache

* Use the right target

* Update workflows

* Update tests

* This is pretty awful

* Use $HOME

* Make sure to create a cache directory

* Shut up

* Use the root directory for rev-list

* Print status

* Go back to GH_TOKEN and construct it from GITHUB_ACTOR

* Checkout --orphan on clone too

* Add a repository dispatch for this

* Leaner image

* Add workaround for tag issue

* Force it

* Try modularity

* Add .targets.mk pre-entrypoint

* Test out 'pre'

* Add an entrypoint and build .targets.mk up front

* Add an action to create a docker image

Maybe it is faster to use this for actions.

* Use image from dockerhub

* Update features

* Update test to skip tags

* Update docker image location

* Remove useless style rule

* Crontab is hard

* Fixup table styles

* Don't push on a pull_request

* Only build gh-pages on push

* Build images monthly too

* Be more generous about what triggers these

* Fix update process

* Make an empty archive if necessary

* Use GITHUB_TOKEN ahead of GH_TOKEN

* Revert "Make an empty archive if necessary"

This reverts commit 1e1b71be1d50c21d13487ea4f43b5e4ae346aff1.

* Use simpler eref

* Make call to GitHub API for default branch.

* A bunch of fixes

I couldn't be bothered splitting these up.

default-branch.py looks locally for a branch name and uses that, either
in origin/HEAD (or $GIT_REMOTE/HEAD) or just HEAD.

It also requires a token if it is going to use the API.

The token variables are rationalized a little.  GITHUB_PUSH_TOKEN is
what is used for pushing.  GITHUB_API_TOKEN is what is used for things
that need the API.

id.mk uses the token to get the default branch, it then exports that for
use by scripts.

I've also updated some of the documentation.

* Remove no-site-packages from virtualenv command

* Bring rfc2629-no-doctype.xslt into lib

Fixes #235

* Update main.mk

* Update main.mk

* Use a tab ffs Mark

* Normalize log() shape

* Suppress silly warning

* Strip rfc-local.css

* Remove extra padding after lists

* Don't right-pad captions

There are a few narrow tables with large captions that will benefit from
this.

* Integrate with Circle's test UI

* Producing HTML is faster

* Fix missing dependency

* Update docker images to work around xml2rfc problems

* Add mmark to images

* Quoting

* A few kB for some sanity

* Fix preprocessing typo

* Disable smart quotes

* More tips for kramdown.

* rough in build-index-md

* control markdown index with MD_INDEX

* add header

* Try using GitHub Packages

* Point to the right file

* Use GitHub Container Registry

* Update action to use branch

* Include GitHub Packages also

* Spell it right

* Use repository owner instead

* GitHub Packages boo

* Use latest label

* Allow for markdown-formatted index

* Use drafts_source more

* Add spellcheck target; not using languagetool

* Fixup drafts_source

* Make HTML files properly for pre-commit

* Remove build-index-md.sh

* Add documentation on linting

This is rough.

* Or try to detect environment...

* Add build to publish.yml, updating linting notes

* Only match draft-*.txt and rfc*.txt

* tabs are a curse

* Lint first

* Don't look at HEAD for the default branch except as a last resort

* Use a bearer token perhaps

* Add a token input

* Use the token parameter

* Damned quoting

* No need for a header on the first set

* Back out whitespace lint

* Remove errant 'let'

* Someone insists that .py is helpful

* Remove raw, which no longer works

* Page numbers are the past

This causes drafts to be rendered in the same way as RFCs, by default.
It opens up the possibility of using tools *other than* rfcdiff for
examining changes between versions.  As textual output is often easier
to handle than HTML when it comes to diff tools, this is a strict
improvement.

This will mean that submissions look different, if the text versions are
submitted.  That is unlikely.  What is more likely is that the submitted
version will be paginated, as that is the xml2rfc default.  This means
that diffs for editors copy (and branches) with the submitted versions
will look worse.  rfcdiff can handle that well enough; other tools
won't be any more usable for that case than they were before.

Overall, this seems to be a strict win.  There might be complaints (I
remain impressed and horrified, but unsurprised, at the reaction to the
lack of page numbers in published RFCs).  That's fine.  I believe that
most people using this toolchain use the HTML rendering primarily, as
that remains more accessible than text.

I am NOT adding a configuration option to disable this, but might be
convinced to.

Closes #251.

* Refinements

* Check for bad tags

This prints a warning when tags aren't good.  I've also modified the
pre-push hook to match.  Of course, hooks are nasty so I don't install
them automatically `make update` does that.

Closes #253.

* Make it black

* The whole thing with .targets.mk is whacky

When you try to upload, make is unable to resolve dependencies that
depend on .targets.mk.  This causes `make upload` to hit the fallback
too easily.  This works around that manually in the most hacky way
imaginable.

Includes a fix to CI for black.

* Downgrade default-branch failure to a warning

* Switch the locale in images

Closes #254.

* Add --name-only

* Add a lint for the default branch

Unfortunately, when you newly create a repository, it doesn't get a
reference for refs/remotes/origin/HEAD.  We should lint for that and
then provide a tool to fix it.

* Third-level headings were too small

* Increase line height; add explicit padding for figures

* easy to miss, hard to unsee

* Build twice a month instead

* Fix a11y for pilcrows

This makes them visible always, but not emphasized so that they don't
distract sighted readers.  This means that screen readers will have this
bump always, but the feedback I received indicated that this was
something that people would prefer to have access to, even if it adds a
little extra noise.

* Remove side-scroll hack for wide tables

The use of inline-block breaks some a11y tools.  The right answer for
overflowing tables is to not overflow.  That is, include less
unwrappable text.

* Dark mode

* Darker

* Fix scrollbars

* Update what setup does

* Switch to using alpine for smaller images

* Use the tag

* Test alternative build structure

...with local registry for testing

* Don't need load with a local registry

* Need driver-opts argument to docker setup

* Try this setup

* Populate the tags properly

* Use the right registry

* Add math/Dockerfile as requested in #259

* kramdown doesn't need more now

* Add math image

* Build math image

Co-authored-by: Martin Thomson <mt@lowentropy.net>
Co-authored-by: Brendan McMillion <brendan@cloudflare.com>
Co-authored-by: Carrick Bartle <cbartle@apple.com>
Co-authored-by: Mark Nottingham <mnot@mnot.net>
Co-authored-by: Mike Bishop <mbishop@evequefou.be>
Co-authored-by: Lars Eggert <lars@eggert.org>